+++ CHANGELOG.md
@@ -17,6 +17,7 @@ but the list below contains much more technical details.
 - CHANGELOG file to keep a full list of changes
 - Save and restore the window state by default
 - Allow specific window properties to be kept or ignored (position, size, maximized state)
+- Press "Shift with t" to open a new tab with the current url prefilled in nav mode
 
 ### Changed